+++ src/components/sections/person/connections.tsx
@@ -18,8 +18,17 @@ import {
   toaster,
   ConfirmButton,
   CancelButton,
+  PopoverRoot,
+  PopoverTrigger,
+  PopoverContent,
+  PopoverArrow,
+  PopoverBody,
+  EditButton,
+  Input,
+  Textarea,
+  EntityDetail,
 } from '@/components'
-import { Person } from '@/gql/graphql'
+import { Person, PersonConnectionsRelationship } from '@/gql/graphql'
 import { useMutation, useQuery } from '@apollo/client'
 import {
   DialogBackdrop,
@@ -27,13 +36,18 @@ import {
   GridItem,
   HStack,
   Spacer,
+  Text,
   VStack,
 } from '@chakra-ui/react'
 import React, { useRef, useState } from 'react'
 import { useForm } from 'react-hook-form'
 
 export default function PersonConnections({ person }: { person: Person }) {
-  const [connections, setConnections] = useState<Person[]>(person.connectedTo)
+  const [edges, setEdges] = useState(person.connectionsConnection.edges)
+  const [openConnectionProps, setOpenConnectionProps] = useState(false)
+  const [selectedConnection, setSelectedConnection] = useState<Person | null>(
+    null
+  )
   const [open, setOpen] = useState(false)
   const { data, loading, error } = useQuery(GET_ALL_PEOPLE)
   const [UpdatePerson] = useMutation(UPDATE_PERSON_MUTATION)
@@ -56,16 +70,29 @@ export default function PersonConnections({ person }: { person: Person }) {
     formState: { isSubmitting, errors },
   } = useForm<FormData>({
     defaultValues: {
-      connections: person.connectedTo?.map((connection) => connection.id) ?? [],
+      connections: person.connections?.map((connection) => connection.id) ?? [],
+    },
+  })
+
+  const {
+    handleSubmit: handleSubmitConnectionProps,
+    control: controlConnectionProps,
+    setValue,
+    formState: {
+      isSubmitting: isSubmittingConnectionProps,
+      errors: errorsConnectionProps,
+    },
+  } = useForm({
+    defaultValues: {
+      why: '',
+      interests: '',
     },
   })
 
   const onSubmit = async (data: FormData) => {
     try {
       // find connections to disconnect and connect
-      const initialConnections = person.connectedTo.map(
-        (connection) => connection.id
-      )
+      const initialConnections = edges.map((edge) => edge.node.id)
       const toDisconnect = initialConnections.filter(
         (connection) => !data.connections.includes(connection)
       )
@@ -93,8 +120,9 @@ export default function PersonConnections({ person }: { person: Person }) {
         throw new Error('No data returned')
       }
 
-      setConnections(
-        response.data.updatePeople.people[0].connectedTo as Person[]
+      setEdges(
+        response.data.updatePeople.people[0].connectionsConnection
+          .edges as PersonConnectionsRelationship[]
       )
       setOpen(false)
 
@@ -111,6 +139,57 @@ export default function PersonConnections({ person }: { person: Person }) {
     }
   }
 
+  type ConnectionPropsData = {
+    why: string
+    interests: string
+  }
+
+  const onSubmitConnectionProps = async (data: ConnectionPropsData) => {
+    try {
+      const response = await UpdatePerson({
+        variables: {
+          where: {
+            id_EQ: person.id,
+          },
+          update: {
+            connections: [
+              {
+                where: {
+                  node: {
+                    id_EQ: selectedConnection?.id,
+                  },
+                },
+                update: {
+                  edge: {
+                    why_SET: data.why,
+                    interests_SET: data.interests,
+                  },
+                },
+              },
+            ],
+          },
+        },
+      })
+
+      if (!response.data) {
+        throw new Error('No data returned')
+      }
+
+      setEdges(
+        response.data.updatePeople.people[0].connectionsConnection
+          .edges as PersonConnectionsRelationship[]
+      )
+      setOpenConnectionProps(false)
+
+      toaster.success({
+        title: 'Updated Connection Properties',
+        description: 'The connection properties have been updated',
+      })
+    } catch (error) {
+      console.error(error)
+    }
+  }
+
   return (
     <ApolloWrapper data={data} loading={loading} error={error}>
       <VStack bg={'bg'} p={4} borderRadius={'2xl'} boxShadow={'xs'}>
@@ -121,7 +200,7 @@ export default function PersonConnections({ person }: { person: Person }) {
         >
           <DialogBackdrop />
           {/* <DialogTrigger asChild> */}
-          {person.connectedTo.length > 0 && (
+          {edges.length > 0 && (
             <HStack width="100%" justifyContent="space-between">
               <Spacer />
               <Button size="sm" variant="surface" onClick={() => setOpen(true)}>
@@ -160,7 +239,7 @@ export default function PersonConnections({ person }: { person: Person }) {
           </form>
         </DialogRoot>
 
-        {person.connectedTo.length === 0 && (
+        {edges.length === 0 && (
           <EmptyState
             title="No Connections"
             description="Click here to add some"
@@ -180,16 +259,108 @@ export default function PersonConnections({ person }: { person: Person }) {
           gap={6}
           width="100%"
         >
-          {connections.map((person) => (
-            <GridItem key={person.id}>
+          {edges.map((edge) => (
+            <GridItem key={edge.node.id} pb={5}>
               <PersonCard
-                id={person.id}
-                photo={person.photo ?? null}
-                name={person.name}
-              />
+                id={edge.node.id}
+                photo={edge.node.photo ?? null}
+                name={edge.node.name}
+              />{' '}
+              <HStack>
+                <VStack>
+                  <PopoverRoot
+                    portalled
+                    positioning={{ placement: 'bottom-end' }}
+                  >
+                    <PopoverTrigger asChild>
+                      <Button
+                        fontSize="x-small"
+                        size="xs"
+                        variant="ghost"
+                        _hover={{
+                          bgColor: 'transparent',
+                        }}
+                      >
+                        {`Connection Details`}
+                      </Button>
+                    </PopoverTrigger>
+                    <PopoverContent>
+                      <PopoverArrow />
+                      <PopoverBody>
+                        <EntityDetail
+                          title="Why"
+                          details={edge.properties.why ?? ''}
+                        />
+                        <EntityDetail
+                          title="Interests"
+                          details={edge.properties.interests ?? ''}
+                        />
+                      </PopoverBody>
+                    </PopoverContent>
+                  </PopoverRoot>
+                </VStack>
+
+                <Spacer />
+                <EditButton
+                  variant="ghost"
+                  size="2xs"
+                  iconOnly
+                  onClick={() => {
+                    setValue('why', edge.properties.why ?? '')
+                    setValue('interests', edge.properties.interests ?? '')
+                    setSelectedConnection(edge.node)
+                    setOpenConnectionProps(true)
+                  }}
+                />
+              </HStack>
             </GridItem>
           ))}
         </Grid>
+        <DialogRoot
+          placement="center"
+          open={openConnectionProps}
+          onOpenChange={(e) => setOpenConnectionProps(e.open)}
+        >
+          <DialogBackdrop />
+          {/* <DialogTrigger asChild> */}
+          <HStack width="100%" justifyContent="space-between">
+            <Spacer />
+            <Text></Text>
+          </HStack>
+          {/* </DialogTrigger> */}
+          <form onSubmit={handleSubmitConnectionProps(onSubmitConnectionProps)}>
+            <DialogContent ref={contentRef}>
+              <DialogHeader>
+                <DialogTitle>{`${person.name} ${selectedConnection?.name} Details`}</DialogTitle>
+              </DialogHeader>
+              <DialogBody>
+                <Input
+                  name="why"
+                  label="Why"
+                  control={controlConnectionProps}
+                  errors={errorsConnectionProps}
+                />
+
+                <Textarea
+                  name="interests"
+                  label="Interests"
+                  control={controlConnectionProps}
+                  errors={errorsConnectionProps}
+                />
+              </DialogBody>
+              <DialogFooter>
+                <DialogActionTrigger asChild>
+                  <CancelButton ref={cancelButtonRef} />
+                </DialogActionTrigger>
+                <ConfirmButton
+                  loading={isSubmittingConnectionProps}
+                  onClick={handleSubmitConnectionProps(onSubmitConnectionProps)}
+                />
+              </DialogFooter>
+              <DialogCloseTrigger />
+            </DialogContent>
+          </form>
+        </DialogRoot>
       </VStack>
     </ApolloWrapper>
   )